requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
rdkit==2023.9.5
faiss-cpu==1.7.4
pyrfume==0.20
python-multipart==0.0.6

# Optional: for web interface
streamlit==1.28.0
plotly==5.18.0

# Development
python-dotenv==1.0.0
jupyter==1.0.0